controller: Consider server mappings for FW-rules
The server mappings (DNS, NTP) are considered the (local)
"administrative context", i.E. the configuration that fills the gap
that MUD leaves for local network administration.

They are passed to the firewall configuration service as an
additional paramter and in the case of a controller match (or a
my-controller match where the mapping specifies a URN, of course)
represent the target devices for which rules are created.