feat(shared): Add extended matching data to firewall rules
The conversion from MUD data to FW rules essentially involves:
 - converting the MUD matches into Hosts
 - getting hosts from L3 matches
 - adding remaining matchable L3 data
 - converting matchable L4 data in full (ports no longer separate)

This removes the idiosyncracy of having ports separate from other L4
data, ensuring that ports will only be specified along with either
TCP or UDP.
However, through the addition of the previously missing L3 matching
data (most of the header data), it does add a separation between the
main L3 data (the hosts) and the remaining (extra) L3 data. This
theoretically makes it possible for a rule to name an IPv4-addressed
host while specifying extra IPv6 header matches data, but IMO it is
a minor snag in the design that probably makes more sense to keep
than to change. Alternatively, it's conceivable to factor the DNS
names out of the L3 data as host information and leave the networks
(adresses) with the L3 data, but I don't see the upside.